requirements.txt
python-dotenv~=1.0.0
web3~=6.2.0
requests~=2.28.2
discord.py~=2.2.2